+++ frontend/src/views/employee/EmployeeDetailTabs.vue
@@ -142,7 +142,7 @@ export default {
         { label: 'org_unit', data: 'org_unit' },
         { label: 'association_type', data: 'association_type' },
         { label: 'address_type', data: 'address_type' },
-        { label: 'address_property', data: 'visibility' },
+        { label: 'visibility', data: 'visibility' },
         { label: 'address', data: 'address' }
       ],
       leave: [
@@ -154,12 +154,12 @@ export default {
         { label: 'manager_type', data: 'manager_type' },
         { label: 'manager_level', data: 'manager_level' },
         { label: 'address_type', data: 'address_type' },
-        { label: 'address_property', data: 'visibility' },
+        { label: 'visibility', data: 'visibility' },
         { label: 'address', data: 'address' }
       ],
       address: [
         { label: 'address_type', data: 'address_type' },
-        { label: 'address_property', data: 'visibility' },
+        { label: 'visibility', data: 'visibility' },
         { label: 'address', data: null }
       ],