+++ frontend/src/views/organisation/OrganisationDetailTabs.vue
@@ -143,7 +143,7 @@ export default {
       ],
       address: [
         { label: 'address_type', data: 'address_type' },
-        { label: 'address_property', data: 'visibility' },
+        { label: 'visibility', data: 'visibility' },
         { label: 'address', data: null }
       ],
       engagement: [
@@ -156,7 +156,7 @@ export default {
         { label: 'person', data: 'person' },
         { label: 'association_type', data: 'association_type' },
         { label: 'address_type', data: 'address_type' },
-        { label: 'address_property', data: 'visibility' },
+        { label: 'visibility', data: 'visibility' },
         { label: 'address', data: 'address' },
         { label: 'org_unit', data: 'org_unit' }
       ],
@@ -174,7 +174,7 @@ export default {
         { label: 'manager_type', data: 'manager_type' },
         { label: 'manager_level', data: 'manager_level' },
         { label: 'address_type', data: 'address_type' },
-        { label: 'address_property', data: 'visibility' },
+        { label: 'visibility', data: 'visibility' },
         { label: 'address', data: 'address' }
       ],
       related_unit: [