+++ NEWS.rst
@@ -12,6 +12,11 @@ merely marking manager functions as *vacant*, set ``"vacant": true``
 in the JSON request. Please note that this is the inverse of the
 previous ``terminate_all`` parameter, which no longer has any affect.
 
+Internal changes
+----------------
+
+* #27431: The ``address_property`` facet is now named ``visibility``.
+
 New features
 ------------